🚚 JE-36-chore: Detail, Review 등 나머지 파일 컴포넌트 분리 및 폴더 구조화
