+++ src/components/ActivityDetail/DetailContent/ActivityTabs.tsx
